variant frame reform (with compiler error)
BUGFIX: rules that are matched in sequence (for example all nameless rules) had so far ignored (print) and (break) tags.

A rule variant A to be invoked is "entered", then each subcalled variant is entered and exited, then "exited": either "success" or "failure". You can now break and print on these events.

The (debug) tag has been forbidden because of deprecation.

Rule variants now have to be verified after being parsed from source code.

And indeed, this code does not compile, there is exactly one compiler error, which I can not explain to myself. Was it there before in older commits?