Introduced variable matching
 - introduced --verbose / -v option

Before, variables were only used to be substituted with their value for the result of a rule application. However, this meant that variables could never be invoked anywhere else than in the rule body. This redundancy has now been solved, since variable invocations in the header or parameter header mean that their contents are matched against the input. (see src/test.swirl)